Create Gitlab-ci-yml-skeleton.yml
In this file, we have defined three stages: build, test, and deploy. We have also defined three jobs corresponding to each stage. Each job is run in a specific Docker image and includes a list of commands to execute in the script attribute.

It's important to note that this .gitlab-ci.yml file is simply a basic example, and you can customize it to fit your needs. You can add additional jobs, dependencies between jobs, trigger rules, etc. For more information on the syntax and features available in .gitlab-ci.yml files, refer to the GitLab documentation.